fix the bugs in rule-based baselines, and change the temporal data sorting strategy
