+++ evaluation/scripts/temporal_locomo/modules/utils.py
@@ -148,24 +148,7 @@ def save_evaluation_cases(
     return saved_files
 
 
-def extract_day_evidences(evidence_list, day):
-    """
-    Extract evidences that belong to the given day.
-
-    Evidence items are expected to be strings in the form "D#:pos", e.g., "D1:3".
-
-    Args:
-        evidence_list: List of evidence strings
-        day: Day identifier string such as "D1"
-
-    Returns:
-        set[str]: Set of evidence strings that start with the given day prefix
-    """
-    day_prefix = f"{day}:"
-    return {e for e in evidence_list if isinstance(e, str) and e.startswith(day_prefix)}
-
-
-def compute_can_answer_stats(day_groups):
+def compute_can_answer_stats(day_groups, rounds_to_consider=float("inf")):
     """
     Compute can-answer statistics for each day using the union of all prior evidences.
 
@@ -182,6 +165,9 @@ def compute_can_answer_stats(day_groups):
     Args:
         day_groups: Dict mapping day_id (e.g., "D1") to a list of QA dicts. Each QA
                     dict should contain an "evidence" field that is a list of strings.
+        rounds_to_consider: Number of previous rounds to consider for evidence accumulation.
+                           Default is infinity (all previous rounds).
+                           Set to 1 to only consider the immediately preceding round.
 
     Returns:
         dict: Mapping day_id -> {"can_answer_count": int, "total": int, "ratio": float}
@@ -189,21 +175,122 @@ def compute_can_answer_stats(day_groups):
     results = {}
     for day, qa_list in day_groups.items():
         seen = set()
+        # Keep track of evidence history for limited rounds
+        evidence_history = []
         can_answer = 0
         total = max(len(qa_list) - 1, 0)
+        rounds_count = 0
         for idx, qa in enumerate(qa_list):
-            cur = extract_day_evidences(qa.get("evidence", []), day)
+            cur = set(qa.get("evidence", []))
+            rounds_count += 1
+
             if idx == 0:
                 # Seed seen evidences with the first QA but do not count it
-                seen |= cur
+                evidence_history.append(cur)
+                seen = set().union(*evidence_history)
                 continue
+
+            # Check if current evidence is subset of accumulated evidence
             if cur and cur.issubset(seen):
                 can_answer += 1
-            # Accumulate all previously asked evidences (union)
-            seen |= cur
+
+            # Add current evidence to history
+            evidence_history.append(cur)
+
+            # Limit history to specified number of rounds
+            if rounds_count > rounds_to_consider:
+                evidence_history.pop(0)
+
+            # Recalculate seen as union of evidence_history
+            seen = set().union(*evidence_history)
+
         results[day] = {
             "can_answer_count": can_answer,
             "total": total,
             "ratio": (can_answer / total) if total else 0.0,
         }
     return results
+
+
+def compute_can_answer_count_by_pre_evidences(
+    temporal_locomo_data, num_of_users, stats_dir=None, rounds_to_consider=float("inf")
+):
+    """
+    Compute can-answer statistics per day for each conversation using the
+    union of all previously asked evidences within the same day.
+
+    Args:
+        temporal_locomo_data: The temporal locomo data containing conversations
+        num_of_users: Number of users/conversations to process
+        stats_dir: Directory to save statistics (optional)
+        rounds_to_consider: Number of previous rounds to consider for evidence accumulation.
+                           Default is infinity (all previous rounds).
+                           Set to 1 to only consider the immediately preceding round.
+
+    Returns:
+        dict: Mapping conversation_id -> per-day stats as produced by compute_can_answer_stats
+    """
+    all_conversations_stats = {}
+    for conv_idx in range(num_of_users):
+        temporal_conv = temporal_locomo_data[conv_idx]
+        conversation_id = temporal_conv["conversation_id"]
+
+        # Build day -> qa_pairs mapping
+        day_groups = {}
+        for day_id, day_data in temporal_conv.get("days", {}).items():
+            day_groups[day_id] = day_data.get("qa_pairs", [])
+
+        # Use shared utility to compute stats with correct accumulation logic
+        per_day_stats = compute_can_answer_stats(day_groups, rounds_to_consider)
+        all_conversations_stats[conversation_id] = per_day_stats
+
+    # Build per-conversation summaries and overall summary
+    per_conversation_summaries = {}
+    overall_can = 0
+    overall_total = 0
+    for conv_id, day_stats in all_conversations_stats.items():
+        conv_can = 0
+        conv_total = 0
+        for _day, stats in day_stats.items():
+            conv_can += int(stats.get("can_answer_count", 0))
+            conv_total += int(stats.get("total", 0))
+        conv_ratio = (conv_can / conv_total) if conv_total else 0.0
+        per_conversation_summaries[conv_id] = {
+            "can_answer_count": conv_can,
+            "total": conv_total,
+            "ratio": conv_ratio,
+        }
+        overall_can += conv_can
+        overall_total += conv_total
+
+    overall_summary = {
+        "can_answer_count": overall_can,
+        "total": overall_total,
+        "ratio": (overall_can / overall_total) if overall_total else 0.0,
+    }
+
+    # Add rounds information to the result
+    result_payload = {
+        "per_conversation_summary": per_conversation_summaries,
+        "overall_summary": overall_summary,
+        "rounds_considered": rounds_to_consider if rounds_to_consider != float("inf") else "all",
+    }
+
+    # Print results
+    print("\nComputed can-answer-by-pre-evidences stats:")
+    print(
+        f"Rounds considered: {rounds_to_consider if rounds_to_consider != float('inf') else 'all'}"
+    )
+    print(json.dumps(result_payload, indent=2, ensure_ascii=False))
+
+    # Save results if stats_dir is provided
+    if stats_dir:
+        output_path = (
+            stats_dir
+            / f"evidences_rounds_{rounds_to_consider if rounds_to_consider != float('inf') else 'all'}.json"
+        )
+        with open(output_path, "w", encoding="utf-8") as fw:
+            json.dump(result_payload, fw, indent=2, ensure_ascii=False)
+        print(f"Saved stats to {output_path}")
+
+    return result_payload
