+++ evaluation/scripts/temporal_locomo/temporal_locomo_eval.py
@@ -1,6 +1,5 @@
 import argparse
 import asyncio
-import json
 import os
 import sys
 
@@ -11,7 +10,7 @@ from locomo_ingestion import LocomoIngestor
 from locomo_metric import LocomoMetric
 from locomo_processor import LocomoProcessor
 from modules.locomo_eval_module import LocomoEvalModelModules
-from modules.utils import compute_can_answer_stats
+from modules.utils import compute_can_answer_count_by_pre_evidences
 
 from memos.log import get_logger
 
@@ -104,7 +103,7 @@ class TemporalLocomoEval(LocomoEvalModelModules):
         print(f"  - Statistics: {self.stats_path}")
         print("=" * 80)
 
-    def compute_can_answer_count_by_pre_evidences(self):
+    def compute_can_answer_count_by_pre_evidences(self, rounds_to_consider):
         """
         Compute can-answer statistics per day for each conversation using the
         union of all previously asked evidences within the same day.
@@ -112,60 +111,12 @@ class TemporalLocomoEval(LocomoEvalModelModules):
         Returns:
             dict: Mapping conversation_id -> per-day stats as produced by compute_can_answer_stats
         """
-        all_conversations_stats = {}
-        for conv_idx in range(self.num_of_users):
-            temporal_conv = self.temporal_locomo_data[conv_idx]
-            conversation_id = temporal_conv["conversation_id"]
-
-            # Build day -> qa_pairs mapping
-            day_groups = {}
-            for day_id, day_data in temporal_conv.get("days", {}).items():
-                day_groups[day_id] = day_data.get("qa_pairs", [])
-
-            # Use shared utility to compute stats with correct accumulation logic
-            per_day_stats = compute_can_answer_stats(day_groups)
-            all_conversations_stats[conversation_id] = per_day_stats
-        # Build per-conversation summaries and overall summary
-        per_conversation_summaries = {}
-        overall_can = 0
-        overall_total = 0
-        for conv_id, day_stats in all_conversations_stats.items():
-            conv_can = 0
-            conv_total = 0
-            for _day, stats in day_stats.items():
-                conv_can += int(stats.get("can_answer_count", 0))
-                conv_total += int(stats.get("total", 0))
-            conv_ratio = (conv_can / conv_total) if conv_total else 0.0
-            per_conversation_summaries[conv_id] = {
-                "can_answer_count": conv_can,
-                "total": conv_total,
-                "ratio": conv_ratio,
-            }
-            overall_can += conv_can
-            overall_total += conv_total
-
-        overall_summary = {
-            "can_answer_count": overall_can,
-            "total": overall_total,
-            "ratio": (overall_can / overall_total) if overall_total else 0.0,
-        }
-
-        result_payload = {
-            "per_conversation_summary": per_conversation_summaries,
-            "overall_summary": overall_summary,
-        }
-
-        # Print results
-        print("\nComputed can-answer-by-pre-evidences stats:")
-        print(json.dumps(result_payload, indent=2, ensure_ascii=False))
-
-        # Save results
-        output_path = self.stats_dir / "compute_can_answer_count_by_pre_evidences.json"
-        with open(output_path, "w", encoding="utf-8") as fw:
-            json.dump(result_payload, fw, indent=2, ensure_ascii=False)
-        print(f"Saved stats to {output_path}")
-
-        return result_payload
+        return compute_can_answer_count_by_pre_evidences(
+            temporal_locomo_data=self.temporal_locomo_data,
+            num_of_users=self.num_of_users,
+            stats_dir=self.stats_dir,
+            rounds_to_consider=rounds_to_consider,
+        )
 
 
 if __name__ == "__main__":
@@ -184,7 +135,7 @@ if __name__ == "__main__":
         help="Version identifier for saving results (e.g., 1010)",
     )
     parser.add_argument(
-        "--workers", type=int, default=1, help="Number of parallel workers to process users"
+        "--workers", type=int, default=10, help="Number of parallel workers to process users"
     )
     parser.add_argument(
         "--top_k", type=int, default=20, help="Number of results to retrieve in search queries"
@@ -199,3 +150,7 @@ if __name__ == "__main__":
 
     evaluator = TemporalLocomoEval(args=args)
     evaluator.run_eval_pipeline()
+
+    # rule-based baselines
+    evaluator.compute_can_answer_count_by_pre_evidences(rounds_to_consider=float("inf"))
+    evaluator.compute_can_answer_count_by_pre_evidences(rounds_to_consider=1)
